Draft works

+++ Assignment3/Programming/HMMTagger/src/HMM/WordTag.java
@@ -5,32 +5,32 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class WordTag {
-    private Map<String, Integer> tagMap;
+    private Map<String, Double> tagMap;
     private int totalWordCount;
 
     public WordTag () {
         this.tagMap = new HashMap<>();
     }
 
-    public Map<String, Integer> getTagMap () {
+    public Map<String, Double> getTagMap () {
         return this.tagMap;
     }
 
-    public Integer getWordTagCount (String tag) {
+    public Double getWordTagCount (String tag) {
         if (tagMap.containsKey(tag)) {
             return tagMap.get(tag);
         } else {
-            return 0;
+            return 0.0;
         }
     }
 
     public void setTagMap (String tag) {
         totalWordCount ++;
         if (tagMap.containsKey(tag)) {
-            Integer count = tagMap.get(tag);
+            Double count = tagMap.get(tag);
             tagMap.put(tag, count + 1);
         } else {
-            tagMap.put(tag, 1);
+            tagMap.put(tag, 1.0);
         }
     }
 }
